Moving a couple classes.
